requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
blis==0.7.7
Bottleneck==1.3.4
catalogue==2.0.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
cymem==2.0.6
datasets==2.3.2
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
fastai==2.6.3
fastapi==0.78.0
fastbook==0.0.26
fastcore==1.4.4
fastdownload==0.0.6
fastprogress==1.0.2
filelock==3.7.1
fonttools==4.31.2
frozenlist==1.3.0
fsspec==2022.5.0
graphviz==0.20
h11==0.13.0
huggingface-hub==0.7.0
idna==3.3
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.9.2
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.0
langcodes==3.3.0
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.7
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
numexpr==2.8.1
numpy==1.22.3
nvidia-ml-py3==7.352.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
preshed==3.0.6
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.21
pydantic==1.8.2
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
regex==2022.6.2
requests==2.27.1
responses==0.18.0
scikit-learn==1.1.1
scipy==1.8.0
Send2Trash==1.8.0
sentencepiece==0.1.96
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.3.1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
sqlparse==0.4.2
srsly==2.4.3
stack-data==0.2.0
starlette==0.19.1
terminado==0.13.3
testpath==0.6.0
thinc==8.0.17
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.11.0+cpu
torchaudio==0.11.0+cpu
torchvision==0.12.0+cpu
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
transformers==4.20.0
typer==0.4.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
xxhash==3.0.0
yarl==1.7.2
